Feature protect

+++ services/app/Http/Controllers/RosterApiController.php
@@ -7,18 +7,39 @@ use App\Models\Employee;
 
 class RosterApiController extends Controller
 {
-    public function index()
+    public function getRoster(Request $request)
     {
+        // Phân quyền xem danh sách nhân viên
+        $role = $request->get('user_roles');
+
+        if (!in_array('manager', $role)) {
+            return response()->json(['message' => 'Lỗi 403: Tính làm hacker hay gì!'], 403);
+        }
+
         return response()->json(Employee::orderBy('created_at', 'desc')->get());
     }
 
-    public function show($id)
+    public function getEmployee(Request $request, $id)
     {
+        // Phân quyền xem chi tiết nhân viên
+        $role = $request->get('user_roles');
+
+        if (!in_array('manager', $role)) {
+            return response()->json(['message' => 'Lỗi 403: Tính làm hacker hay gì!'], 403);
+        }
+
         return response()->json(Employee::find($id));
     }
 
-    public function store(Request $request)
+    public function storeEmployee(Request $request)
     {
+        // Phân quyền thêm nhân viên
+        $role = $request->get('user_roles');
+
+        if (!in_array('manager', $role)) {
+            return response()->json(['message' => 'Lỗi 403: Tính làm hacker hay gì!'], 403);
+        }
+
         try {
             $input_data = $request->all();
             if(!isset($input_data['TRANGTHAI'])) $input_data['TRANGTHAI'] = 1; 
@@ -30,8 +51,14 @@ class RosterApiController extends Controller
         }
     }
 
-    public function update(Request $request, $id)
+    public function updateEmployee(Request $request, $id)
     {
+        // Phân quyền sửa nhân viên
+        $role = $request->get('user_roles');
+
+        if (!in_array('manager', $role)) {
+            return response()->json(['message' => 'Lỗi 403: Tính làm hacker hay gì!'], 403);
+        }
         try {
             $employee = Employee::find($id);
             if($employee) {
@@ -44,10 +71,18 @@ class RosterApiController extends Controller
         }
     }
 
-    public function destroy($id)
+    public function destroyEmployee(Request $request, $id)
     {
+        // Phân quyền xóa nhân viên
+        $role = $request->get('user_roles');
+
+        if (!in_array('manager', $role)) {
+            return response()->json(['message' => 'Lỗi 403: Tính làm hacker hay gì!'], 403);
+        }
+
         $employee = Employee::find($id);
         if ($employee) {
+            \App\Models\Payroll::where('MANV', $id)->delete();
             $employee->delete();
             return response()->json(['success' => true]);
         }
